==> 1_data_generation.Rmd <==
---
title: "Data Generation Process"
output: html_document
editor_options: 
  chunk_output_type: console
---

Variables in generated datasets

* W: True variable
* X: Observed variable
* Y: Background variables
* Z: Audit inclusion

Number of conditions

* 3 relationship between 'true' variable W and 'observed' variable X
* 3 relationship between 'true' variable W and profiles Y
* 3 relationship between 'true' variable W and 'already audited' Z 


```{r nconditions}

options(scipen = )

set.seed(123)
x_ncon = c(1)
y_ncon = c(1)
z_ncon = c(1, 2, 3, 4)
```

Combination of all specified conditions
```{r combinations}
conditions = list(x_ncon = x_ncon, y_ncon = y_ncon, z_ncon = z_ncon)
xyzw_con = expand.grid(conditions)
```

Define each condition using the joint probabilities in the bivariate relationship

For the XW relationship: 

* perfect relation, probability of X=1 when W=1 is 1

```{r listXW}
WX = list()
WX[[1]] = matrix(c(1/3, 0, 0, 
                   0, 1/3, 0,
                   0, 0, 1/3), nrow = 3, ncol = 3, byrow = TRUE)
``` 

For the WY relationship:

* probability of X=1 when Y=1 is .8

```{r listXY}
WY = list()
WY[[1]] = matrix(c(.8/3, .1/3, .1/3, 
                   .1/3, .8/3, .1/3,
                   .1/3, .1/3, .8/3), nrow = 3, ncol = 3, byrow = TRUE)
```

For the WZ relationship:

* probability of Z=1 is .1 and probability of Z=2 is .9, probability of Y=1 when Z=1 is .1/3 and when Z=2 is .3 and is equal for Y=2 and Y=3
* probability of Z=1 is .1 and probability of Z=2 is .9, probability of Y=1 when Z=1 is .06 when Z=2 is .3, Y=2 when Z=1 is .03 and Y=3 when Z=1 is .01
* probability of Z=1 is .3 and probability of Z=2 is .7, probability of Y=1 when Z=1 and when Z=2 are both 1/6 and are equal for Y=2 and Y=3
* probability of Z=1 is .3 and probability of Z=2 is .7, probability of Y=1 when Z=1 is .3 and when Z=2 1/6, Y=2 when Z=1 is .15 and Y=3 when Z=1 is .05

```{r list YZ}
WZ = list()
WZ[[1]] = matrix(c(.01/3, .97/3, 
                   .01/3, .97/3,
                   .01/3, .97/3), nrow = 3, ncol = 2, byrow = TRUE)
WZ[[2]] = matrix(c(.012, .97/3, 
                   .010, .97/3,
                   .008, .97/3), nrow = 3, ncol = 2, byrow = TRUE)
WZ[[3]] = matrix(c(.015, .97/3, 
                   .010, .97/3,
                   .005, .97/3), nrow = 3, ncol = 2, byrow = TRUE)
WZ[[4]] = matrix(c(.018, .97/3, 
                   .010, .97/3,
                   .002, .97/3), nrow = 3, ncol = 2, byrow = TRUE)
```

List the number of categories for X, Y, W and Z and create an expanded grid
```{r allcomb}
X = c(1,2,3)
Y = c(1,2,3)
W = c(1,2,3)
Z = c(1,2)

variables = list(X = X, Y = Y, W = W, Z=Z)
XYWZ = expand.grid(variables)
```

Loop over all conditions (combinations of different relationships between XYWZ) and store the generated datasets in a list

```{r genloop}
generated_datasets = list()
generated_distributions = list()

# loop over all conditions (combinations of different relations between X Y W Z )
for (i in 1:nrow(xyzw_con)){
  
  for (j in 1:nrow(XYWZ)){
    # if X=... and W=... find the corresponding probability in the matrix WX
    if(XYWZ[j,'W'] == 1 & XYWZ[j,'X'] == 1){XYWZ[j,'WX'] = WX[[as.numeric(paste0(xyzw_con[i,1]))]][1,1]}
    if(XYWZ[j,'W'] == 1 & XYWZ[j,'X'] == 2){XYWZ[j,'WX'] = WX[[as.numeric(paste0(xyzw_con[i,1]))]][1,2]}
    if(XYWZ[j,'W'] == 1 & XYWZ[j,'X'] == 3){XYWZ[j,'WX'] = WX[[as.numeric(paste0(xyzw_con[i,1]))]][1,3]}
    if(XYWZ[j,'W'] == 2 & XYWZ[j,'X'] == 1){XYWZ[j,'WX'] = WX[[as.numeric(paste0(xyzw_con[i,1]))]][2,1]}
    if(XYWZ[j,'W'] == 2 & XYWZ[j,'X'] == 2){XYWZ[j,'WX'] = WX[[as.numeric(paste0(xyzw_con[i,1]))]][2,2]}
    if(XYWZ[j,'W'] == 2 & XYWZ[j,'X'] == 3){XYWZ[j,'WX'] = WX[[as.numeric(paste0(xyzw_con[i,1]))]][2,3]}
    if(XYWZ[j,'W'] == 3 & XYWZ[j,'X'] == 1){XYWZ[j,'WX'] = WX[[as.numeric(paste0(xyzw_con[i,1]))]][3,1]}
    if(XYWZ[j,'W'] == 3 & XYWZ[j,'X'] == 2){XYWZ[j,'WX'] = WX[[as.numeric(paste0(xyzw_con[i,1]))]][3,2]}
    if(XYWZ[j,'W'] == 3 & XYWZ[j,'X'] == 3){XYWZ[j,'WX'] = WX[[as.numeric(paste0(xyzw_con[i,1]))]][3,3]}
    # if X=... and Y=... find the corresponding probability in the matrix WY
    if(XYWZ[j,'W'] == 1 & XYWZ[j,'Y'] == 1){XYWZ[j,'WY'] = WY[[as.numeric(paste0(xyzw_con[i,2]))]][1,1]}
    if(XYWZ[j,'W'] == 1 & XYWZ[j,'Y'] == 2){XYWZ[j,'WY'] = WY[[as.numeric(paste0(xyzw_con[i,2]))]][1,2]}
    if(XYWZ[j,'W'] == 1 & XYWZ[j,'Y'] == 3){XYWZ[j,'WY'] = WY[[as.numeric(paste0(xyzw_con[i,2]))]][1,3]}
    if(XYWZ[j,'W'] == 2 & XYWZ[j,'Y'] == 1){XYWZ[j,'WY'] = WY[[as.numeric(paste0(xyzw_con[i,2]))]][2,1]}
    if(XYWZ[j,'W'] == 2 & XYWZ[j,'Y'] == 2){XYWZ[j,'WY'] = WY[[as.numeric(paste0(xyzw_con[i,2]))]][2,2]}
    if(XYWZ[j,'W'] == 2 & XYWZ[j,'Y'] == 3){XYWZ[j,'WY'] = WY[[as.numeric(paste0(xyzw_con[i,2]))]][2,3]}
    if(XYWZ[j,'W'] == 3 & XYWZ[j,'Y'] == 1){XYWZ[j,'WY'] = WY[[as.numeric(paste0(xyzw_con[i,2]))]][3,1]}
    if(XYWZ[j,'W'] == 3 & XYWZ[j,'Y'] == 2){XYWZ[j,'WY'] = WY[[as.numeric(paste0(xyzw_con[i,2]))]][3,2]}
    if(XYWZ[j,'W'] == 3 & XYWZ[j,'Y'] == 3){XYWZ[j,'WY'] = WY[[as.numeric(paste0(xyzw_con[i,2]))]][3,3]}
    # If Y=.. and Z=... find the corresponding probability in the matrix WZ
    if(XYWZ[j,'W'] == 1 & XYWZ[j,'Z'] == 1){XYWZ[j,'WZ'] = WZ[[as.numeric(paste0(xyzw_con[i,3]))]][1,1]}
    if(XYWZ[j,'W'] == 1 & XYWZ[j,'Z'] == 2){XYWZ[j,'WZ'] = WZ[[as.numeric(paste0(xyzw_con[i,3]))]][1,2]}
    if(XYWZ[j,'W'] == 2 & XYWZ[j,'Z'] == 1){XYWZ[j,'WZ'] = WZ[[as.numeric(paste0(xyzw_con[i,3]))]][2,1]}
    if(XYWZ[j,'W'] == 2 & XYWZ[j,'Z'] == 2){XYWZ[j,'WZ'] = WZ[[as.numeric(paste0(xyzw_con[i,3]))]][2,2]}
    if(XYWZ[j,'W'] == 3 & XYWZ[j,'Z'] == 1){XYWZ[j,'WZ'] = WZ[[as.numeric(paste0(xyzw_con[i,3]))]][3,1]}
    if(XYWZ[j,'W'] == 3 & XYWZ[j,'Z'] == 2){XYWZ[j,'WZ'] = WZ[[as.numeric(paste0(xyzw_con[i,3]))]][3,2]}
  
  }
  
  # Assuming independence (relation between XW is independent of relation XY and YZ, we multiply the 3 probabilities
  # multiply the number by 9 to make all final probabilities sum to one)
  XYWZ[,'prob']                = (XYWZ[,'WX']*XYWZ[,'WY']*XYWZ[,'WZ'])*9
  generated_distributions[[i]] = XYWZ
  
  # now generating 1 dataset per simulation condition
  generated_datasets[[i]] = cbind(XYWZ[,1:4], rmultinom(1000, size = 10000, prob = XYWZ[,'prob']))

}
```

Save generated datasets
```{r savedata}
save(generated_distributions, file = "distributions.RData")
save(generated_datasets,      file = "datasets.RData")
```
